Registro de usuarios conexion db

+++ database/migrations/2016_12_12_230242_create_network_accounts_table.php
@@ -16,7 +16,7 @@ class CreateNetworkAccountsTable extends Migration
         Schema::create('network_accounts', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('user_id')->unsigned();
-            $table->integer('provider_id');
+            $table->bigInteger('provider_id');
             $table->string('provider');
             $table->timestamps();
         });
